renaming worker to common
