+++ core-common/.gitignore
