+++ core-common/proguard-rules.pro
