+++ core-common/src/main/java/com/centroi/alsuper/core/common/Constants.kt
@@ -1,0 +1,3 @@
+package com.centroi.alsuper.core.common
+
+const val WORK_TIME = 15
