+++ core-common/src/main/java/com/centroi/alsuper/core/common/ForceKaptModule.kt
@@ -1,0 +1,13 @@
+package com.centroi.alsuper.core.common
+
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+
+@Module
+@InstallIn(SingletonComponent::class)
+object ForceKaptModule {
+    @Provides
+    fun force(): Boolean = true
+}
