+++ core-common/src/main/java/com/centroi/alsuper/core/common/location/ProviderModule.kt
@@ -1,0 +1,16 @@
+package com.centroi.alsuper.core.common.location
+
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+
+@Module
+@InstallIn(SingletonComponent::class)
+interface ProviderModule {
+
+    @Binds
+    fun bindLocationProvider(
+        impl: DefaultLocationProvider
+    ): LocationProvider
+}
