--- core-worker/build.gradle.kts
@@ -1,119 +1,0 @@
-@Suppress("DSL_SCOPE_VIOLATION") // Remove when fixed https://youtrack.jetbrains.com/issue/KTIJ-19369
-plugins {
-    alias(libs.plugins.android.library)
-    alias(libs.plugins.kotlin.android)
-    alias(libs.plugins.compose.compiler)
-    alias(libs.plugins.kotlin.kapt)
-    alias(libs.plugins.detekt)
-    alias(libs.plugins.hilt.gradle)
-    id("jacoco")
-}
-
-android {
-    namespace = "com.centroi.alsuper.core.worker"
-    compileSdk = 35
-
-    defaultConfig {
-        minSdk = 23
-
-        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-        consumerProguardFiles("consumer-rules.pro")
-    }
-
-    kapt {
-        correctErrorTypes = true
-    }
-
-    buildFeatures {
-        compose = true
-        aidl = false
-        buildConfig = false
-        renderScript = false
-        shaders = false
-    }
-
-    compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_17
-        targetCompatibility = JavaVersion.VERSION_17
-    }
-    kotlinOptions {
-        jvmTarget = "17"
-    }
-}
-
-dependencies {
-    implementation(project(":core-data"))
-    implementation(project(":core-database"))
-    implementation(libs.play.services.location)
-    implementation(libs.androidx.hilt.common)
-    implementation(libs.androidx.work.runtime.ktx)
-    implementation(libs.androidx.work.testing)
-    val composeBom = platform(libs.androidx.compose.bom)
-    implementation(composeBom)
-    androidTestImplementation(composeBom)
-
-    // Core Android dependencies
-    implementation(libs.androidx.core.ktx)
-
-    // Hilt Dependency Injection
-    implementation(libs.hilt.android)
-    implementation(libs.androidx.ui.test.junit4.android)
-    kapt(libs.hilt.compiler)
-
-    // Compose
-    implementation(libs.androidx.compose.ui)
-    implementation(libs.androidx.compose.ui.tooling.preview)
-    implementation(libs.androidx.compose.material3)
-    implementation(libs.androidx.material3.adaptive.navigation.suite.android)
-    implementation(libs.androidx.navigation.compose)
-    implementation(libs.androidx.hilt.navigation.compose)
-
-    // Mockito dependencies
-    testImplementation(libs.mockito.core)
-    testImplementation(libs.mockito.kotlin)
-    testImplementation(libs.mockito.inline) // Enables mocking final classes
-
-    // Tooling
-    debugImplementation(libs.androidx.compose.ui.tooling)
-}
-
-tasks.register<JacocoReport>("jacocoTestReport") {
-    dependsOn("testDebugUnitTest")
-
-    reports {
-        xml.required.set(true)
-        html.required.set(true)
-    }
-
-    val fileFilter = listOf(
-        "**/R.class",
-        "**/R\$*.class",
-        "**/BuildConfig.*",
-        "**/Manifest*.*",
-        "**/*Activity*.*",
-        "**/*Fragment*.*",
-        "**/*Screen*.*" // Exclude UI components (optional)
-    )
-
-    val debugTree = fileTree("${buildDir}/intermediates/javac/debug") {
-        exclude(fileFilter)
-    }
-    val kotlinDebugTree = fileTree("${buildDir}/tmp/kotlin-classes/debug") {
-        exclude(fileFilter)
-    }
-
-    sourceDirectories.setFrom(files("src/main/java"))
-    classDirectories.setFrom(files(debugTree, kotlinDebugTree))
-    executionData.setFrom(files("${buildDir}/jacoco/testDebugUnitTest.exec"))
-}
-
-tasks.register<JacocoCoverageVerification>("jacocoCoverageVerification") {
-    dependsOn("jacocoTestReport")
-    violationRules {
-        rule {
-            limit {
-                minimum = BigDecimal("0.80") // At least 80% coverage required
-            }
-        }
-    }
-}
