--- core-worker/src/main/java/com/centroi/alsuper/core/worker/Constants.kt
@@ -1,3 +1,0 @@
-package com.centroi.alsuper.core.worker
-
-const val WORK_TIME = 15
