+++ feature-auth/src/main/java/com/centroi/alsuper/feature/auth/screens/RegistrationScreenViewModel.kt
@@ -1,0 +1,9 @@
+package com.centroi.alsuper.feature.auth.screens
+
+import androidx.lifecycle.ViewModel
+
+class RegistrationScreenViewModel() : ViewModel() {
+
+
+
+}
