+++ settings.gradle.kts
@@ -41,6 +41,6 @@ include(":feature-auth")
 include(":feature-fakeapp")
 include(":feature-profile")
 include(":feature-contacts")
-include(":core-worker")
+include(":core-common")
 include(":feature-information")
 include(":feature-chat")
